feat: implement createInSubfolder setting
