+++ editor/src/helpers.ts
@@ -1,6 +1,6 @@
 export function debounce<A = unknown, R = void>(
   fn: (args?: A) => R,
-  ms: number
+  ms: number,
 ): [(args?: A) => Promise<R>, () => void] {
   let timer: ReturnType<typeof setTimeout>;
